refactor(Dto): Change the dto location
